have weather and movies showing up
