+++ modules/weather.js
@@ -1,0 +1,28 @@
+'use strict';
+
+const axios = require('axios');
+
+class Forecast {
+  constructor(day) {
+    this.date = `Date: ${day.valid_date}.`;
+    this.des = `Has ${day.weather.description}`;
+  }
+}
+
+async function getWeather (request, response) {
+  let multiDayWeather = [];
+  try {
+    let getLat = request.query.lat;
+    let getLon = request.query.lon;
+    
+    multiDayWeather = await axios.get(`https://api.weatherbit.io/v2.0/forecast/daily?lat=${getLat}&lon=${getLon}&key=${process.env.WEATHER_API_KEY}`);
+    console.log(multiDayWeather.data.data);
+    
+    response.send(multiDayWeather.data.data.map(day => new Forecast(day)));
+  } catch (error) {
+    console.log(`Error: ${error}`)
+  }
+  
+};
+
+module.exports = getWeather;
