+++ server.js
@@ -22,78 +22,12 @@ const PORT = process.env.PORT;
 // everything above this line is what we need for an express server (or close it)
 console.log('Port Number is: ', PORT)
 
-class Forecast {
-  constructor(day) {
-    this.date = `Date: ${day.valid_date}.`;
-    this.des = `Has ${day.weather.description}`;
-  }
-}
+let getWeather = require('./modules/weather.js');
+let getMovies = require('./modules/movies.js');
 
-class Movies {
-  constructor(movie) {
-    this.imageUrl = `"Image Url": ${movie.backdrop_path}.`;
-    this.overview = `"Overview": ${movie.overview}.`;
-    this.popularity = `"Popularity": ${movie.popularity}.`;
-    this.released = `"Released On": ${movie.release_date}.`;
-    this.title = `"Title": ${movie.title}.`;
-    this.totalVotes = `"Total Votes": ${movie.vote_count}.`;
-    this.votes = `"Votes": ${movie.vote_average}.`;
-  }
-}
+app.get('/weather', getWeather)
 
-app.get('/weather', async (request, response) => {
-  let multiDayWeather = [];
-  try {
-    let searchQuery = request.query.city_name;
-
-    let weatherData = await axios.get(`https://api.weatherbit.io/v2.0/current?city=${searchQuery}&key=${process.env.WEATHER_API_KEY}`);
-
-    // Pull out the lon and lat from this api
-    let lon = weatherData.data.data[0].lon
-    let lat = weatherData.data.data[0].lat
-
-    // Use that lon and lat into the new api
-    multiDayWeather = await axios.get(`https://api.weatherbit.io/v2.0/forecast/daily?lat=${lat}&lon=${lon}&key=${process.env.WEATHER_API_KEY}`);
-
-    response.send(multiDayWeather.data.data.map(day => new Forecast(day)));
-  } catch (error) {
-    console.log(`Error: ${error}`)
-  }
-
-});
-
-app.get('/movies', async (request, response) => {
-  try {
-    let searchQuery = request.query.query;
-
-    // let searchQuery = 'Seattle';
-
-    let movies = await axios.get(`https://api.themoviedb.org/3/search/movie?api_key=${process.env.MOVIE_API_KEY}&query=${searchQuery}`)
-
-
-    console.log(movies.data.results);
-
-    // let movieName =
-
-
-    response.send(movies.data.results.map(movie => new Movies(movie)));
-  } catch (error) {
-    console.log(`Error: ${error}`)
-  }
-})
-
-
-
-
-
-
-
-
-
-
-
-
-
+app.get('/movies', getMovies)
 
 
 // specify what routes our server should be listening for
